# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "A4PM Receitas API",
        "description": "Documentação da API de receitas e usuários",
        "version": "1.0"
    },
    "paths": {
        "/receitas/dashboard": {
            "get": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Lista todas as receitas do usuário autenticado",
                "operationId": "778b3944862dd24b190686dd0a65a696",
                "responses": {
                    "200": {
                        "description": "Lista de receitas retornada com sucesso"
                    }
                }
            }
        },
        "/receitas": {
            "post": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Cria uma nova receita",
                "operationId": "bd75738047a6505675761c40d7872098",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "modo_preparo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "modo_preparo": {
                                        "type": "string"
                                    },
                                    "ingredientes": {
                                        "type": "string"
                                    },
                                    "tempo_preparo_minutos": {
                                        "type": "integer"
                                    },
                                    "porcoes": {
                                        "type": "integer"
                                    },
                                    "id_categorias": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Receita criada com sucesso"
                    },
                    "422": {
                        "description": "Erros de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "nome": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "['Campo nome deve ter no máximo 100 caracteres.']"
                                                    }
                                                },
                                                "id_categorias": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "['Não foi possivel cadastrar a receita com essa categoria selecionada, já que a categoria não existe.']"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/receitas/{id}": {
            "put": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Atualiza uma receita existente",
                "operationId": "238ffb6d440a42029d2f4d485cad0040",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da receita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "modo_preparo": {
                                        "type": "string"
                                    },
                                    "ingredientes": {
                                        "type": "string"
                                    },
                                    "tempo_preparo_minutos": {
                                        "type": "integer"
                                    },
                                    "porcoes": {
                                        "type": "integer"
                                    },
                                    "id_categorias": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receita atualizada com sucesso"
                    },
                    "422": {
                        "description": "Erros de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "nome": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "['Campo nome deve ter no máximo 100 caracteres.']"
                                                    }
                                                },
                                                "id_categorias": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "['Não foi possivel cadastrar a receita com essa categoria selecionada, já que a categoria não existe.']"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Exclui uma receita",
                "operationId": "2ef390e37df948d92ff387726318235a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da receita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Receita excluída com sucesso"
                    }
                }
            }
        },
        "/receitas/{id}/imprimir": {
            "get": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Renderiza a receita para impressão",
                "operationId": "6af759e30c781d02672e007d95794054",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da receita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receita pronta para impressão"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Autentica um usuário",
                "operationId": "16a24e707e2a7bdee257b2d0fab49039",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "usuario123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redireciona para o dashboard após login"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    },
                    "422": {
                        "description": "Erros de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "login": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "['O campo login é obrigatório.', 'O login deve ter no mínimo 2 caracteres.', 'O login deve ter no máximo 100 caracteres.']"
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "['O campo senha é obrigatório.', 'A senha deve deve ter no mínimo 2 caracteres.', 'A senha deve ter no máximo 100 caracteres.']"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cadastro": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Registra um novo usuário",
                "operationId": "8a392851bb81bd52c6ffa5d0ecec9404",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "joaosilva"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senhaSegura123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redireciona para o dashboard após cadastro com sucesso"
                    },
                    "422": {
                        "description": "Erros de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "login": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "['O campo login é obrigatório.', 'O login deve ter no mínimo 2 caracteres.', 'O login deve ter no máximo 100 caracteres.', 'O login deve ser único, ou seja, não deve existir outro usuario com o mesmo login escolhido.']"
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "['O campo senha é obrigatório.', 'A senha deve deve ter no mínimo 2 caracteres.', 'A senha deve ter no máximo 100 caracteres.']"
                                                    }
                                                },
                                                "nome": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "['O campo nome é obrigatório.', 'O nome deve deve ter no mínimo 2 caracteres.', 'O nome deve ter no máximo 100 caracteres.']"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Receitas",
            "description": "Operações relacionadas às receitas"
        },
        {
            "name": "Usuários",
            "description": "Operações de autenticação e cadastro de usuários"
        }
    ]
}